[refactor] Extract StreamMultipleInputProcessor factory method
